+++ daemon/network/message_handler/kill.rs
@@ -4,7 +4,7 @@ use pueue_lib::network::message::*;
 use pueue_lib::state::SharedState;
 
 use super::SENDER_ERR;
-use crate::network::response_helper::{ensure_group_exists, task_response_helper};
+use crate::network::response_helper::{ensure_group_exists, task_action_response_helper};
 
 /// Invoked when calling `pueue kill`.
 /// Forward the kill message to the task handler, which then kills the process.
@@ -23,15 +23,12 @@ pub fn kill(message: KillMessage, sender: &Sender<Message>, state: &SharedState)
 
     if let Some(signal) = message.signal {
         match message.tasks {
-            TaskSelection::TaskIds(task_ids) => {
-                let response = task_response_helper(
-                    "Tasks are being killed",
-                    task_ids,
-                    |task| task.is_running(),
-                    &state,
-                );
-                create_success_message(response)
-            }
+            TaskSelection::TaskIds(task_ids) => task_action_response_helper(
+                "Tasks are being killed",
+                task_ids,
+                |task| task.is_running(),
+                &state,
+            ),
             TaskSelection::Group(group) => create_success_message(format!(
                 "Sending signal {} to all running tasks of group {}.",
                 signal, group
@@ -42,15 +39,12 @@ pub fn kill(message: KillMessage, sender: &Sender<Message>, state: &SharedState)
         }
     } else {
         match message.tasks {
-            TaskSelection::TaskIds(task_ids) => {
-                let response = task_response_helper(
-                    "Tasks are being killed",
-                    task_ids,
-                    |task| task.is_running(),
-                    &state,
-                );
-                create_success_message(response)
-            }
+            TaskSelection::TaskIds(task_ids) => task_action_response_helper(
+                "Tasks are being killed",
+                task_ids,
+                |task| task.is_running(),
+                &state,
+            ),
             TaskSelection::Group(group) => create_success_message(format!(
                 "All tasks of group \"{}\" are being killed.",
                 group