+++ daemon/network/response_helper.rs
@@ -1,6 +1,6 @@
 use std::sync::MutexGuard;
 
-use pueue_lib::network::message::{create_failure_message, Message};
+use pueue_lib::network::message::{create_failure_message, create_success_message, Message};
 use pueue_lib::state::State;
 use pueue_lib::task::Task;
 
@@ -17,12 +17,17 @@ pub fn ensure_group_exists(state: &MutexGuard<State>, group: &str) -> Result<(),
     Ok(())
 }
 
-pub fn task_response_helper<F>(
+/// Compile a response for actions that affect several given tasks.
+/// These actions can sometimes only succeed for a part of the given tasks.
+///
+/// That's why this helper exists, which determines based on a given criterion `filter`
+/// for which tasks the action succeeded and which tasks failed.
+pub fn task_action_response_helper<F>(
     message: &str,
     task_ids: Vec<usize>,
     filter: F,
     state: &MutexGuard<State>,
-) -> String
+) -> Message
 where
     F: Fn(&Task) -> bool,
 {
@@ -39,14 +44,14 @@ pub fn compile_task_response(
     message: &str,
     matching: Vec<usize>,
     mismatching: Vec<usize>,
-) -> String {
+) -> Message {
     let matching: Vec<String> = matching.iter().map(|id| id.to_string()).collect();
     let mismatching: Vec<String> = mismatching.iter().map(|id| id.to_string()).collect();
     let matching_string = matching.join(", ");
 
     // We don't have any mismatching ids, return the simple message.
     if mismatching.is_empty() {
-        return format!("{}: {}", message, matching_string);
+        return create_success_message(format!("{}: {}", message, matching_string));
     }
 
     let mismatched_message = "The command failed for tasks";
@@ -54,12 +59,12 @@ pub fn compile_task_response(
 
     // All given ids are invalid.
     if matching.is_empty() {
-        return format!("{}: {}", mismatched_message, mismatching_string);
+        return create_failure_message(format!("{}: {}", mismatched_message, mismatching_string));
     }
 
     // Some ids were valid, some were invalid.
-    format!(
+    create_success_message(format!(
         "{}: {}\n{}: {}",
         message, matching_string, mismatched_message, mismatching_string
-    )
+    ))
 }